Update server_postgres_MP.py
Data is currently being loaded to a base table called BASE in the format TIMESTAMP | TEXT. This table will need to be further split into the further dimensional tables. I would love to hear if anyone has any ideas about how best to do this. The tables are currently in the data warehouse, but with no data